lint query params
