+++ .custom-gcl.yml
@@ -6,8 +6,8 @@ plugins:
     path: ./tools/fmtpercentv
   - module: "github.com/google/go-github/v88/tools/redundantptr"
     path: ./tools/redundantptr
-  - module: "github.com/google/go-github/v88/tools/requestbody"
-    path: ./tools/requestbody
+  - module: "github.com/google/go-github/v88/tools/paramcheck"
+    path: ./tools/paramcheck
   - module: "github.com/google/go-github/v88/tools/sliceofpointers"
     path: ./tools/sliceofpointers
   - module: "github.com/google/go-github/v88/tools/structfield"
